requirements.txt
openai==0.26.5
toml==0.10.2
pydantic==1.10.4
python-telegram-bot==20.1
SQLAlchemy==2.0.3
loguru==0.6.0
psycopg2==2.9.3
